implement purchase

+++ src/main/java/com/hfsolution/feature/token/repository/TokenRepository.java
@@ -17,11 +17,11 @@ public interface TokenRepository extends JpaRepository<Token, Integer> {
       on t.user.id = u.id\s
       where u.id = :id and (t.expired = false or t.revoked = false)\s
       """)
-  List<Token> findAllValidTokenByUser(Integer id);
+  List<Token> findAllValidTokenByUser(Long id);
 
   @Modifying
   @Transactional
-  void deleteByUserId(Integer userId);
+  void deleteByUserId(Long userId);
 
   Optional<Token> findByToken(String token);
 }
